Juegos Terminado 4-6 Entidades Con Alertas

+++ src/main/java/com/grupo5/webapp/steam/controller/FXController/DesarrolladorControllerFX.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Component;
 import com.grupo5.webapp.steam.model.Desarrollador;
 import com.grupo5.webapp.steam.service.DesarrolladorService;
 import com.grupo5.webapp.steam.system.Main;
+import com.grupo5.webapp.steam.utils.SteamAlert;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -17,6 +18,7 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
@@ -45,15 +47,31 @@ public class DesarrolladorControllerFX implements Initializable{
     }
     public void handleButtonAction(ActionEvent event) {
         if(event.getSource() == btnGuardar){
-            if(tfId.getText().isBlank()){
-                agregarDesarrollador();
-            }else{
-                editarDesarrollador();
+
+            boolean exito;
+            if (tfId.getText().isBlank()) {
+                exito = agregarDesarrollador();
+                if (exito) {
+                    SteamAlert.getInstance().mostrarAlertaInfo(401);
+                    cargarDatos();
+                }else{
+
+                }
+            } else {
+                if (SteamAlert.getInstance().mostrarAlertaConfirmacion(406).get() == ButtonType.OK) {
+                    exito = editarDesarrollador();
+                    if (exito) {
+                        cargarDatos();
+                    } else {
+                    }
+                }
             }
         }else if(event.getSource() == btnLimpiar){
             limpiarForm();
         }else if(event.getSource() == btnEliminar){
-            eliminarDesarrollador();
+             if(SteamAlert.getInstance().mostrarAlertaConfirmacion(405).get() == ButtonType.OK){
+                eliminarDesarrollador();
+            }
         }else if(event.getSource() == btnRegresar){
             stage.indexView();
         }else if (event.getSource() == btnBuscar) {
@@ -61,7 +79,7 @@ public class DesarrolladorControllerFX implements Initializable{
             if (tfBuscar.getText().isBlank()) {
                 cargarDatos();
             } else {
-                tblDesarrolladores.setItems(listarDesarrollador());
+                tblDesarrolladores.getItems().add(buscarDesarrollador());
                 colId.setCellValueFactory(new PropertyValueFactory<Desarrollador,Long>("id"));
                 colNombre.setCellValueFactory(new PropertyValueFactory<Desarrollador,String>("nombre"));
                 colPais.setCellValueFactory(new PropertyValueFactory<Desarrollador,String>("pais"));
@@ -87,12 +105,16 @@ public class DesarrolladorControllerFX implements Initializable{
     }
 
 
-    public void agregarDesarrollador(){
-        Desarrollador desarrollador = new Desarrollador();
-        desarrollador.setNombre(tfNombre.getText());
-        desarrollador.setPais(tfPais.getText());
-        desarrolladorService.guardarDesarrollador(desarrollador);
-        cargarDatos();
+    public Boolean agregarDesarrollador(){
+        try {
+            Desarrollador desarrollador = new Desarrollador();
+            desarrollador.setNombre(tfNombre.getText());
+            desarrollador.setPais(tfPais.getText());
+            return desarrolladorService.guardarDesarrollador(desarrollador);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return false;
+        }
     }
 
     public void cargarFormEditar(){
@@ -103,12 +125,17 @@ public class DesarrolladorControllerFX implements Initializable{
             tfPais.setText(desarrollador.getPais());
         }
     }
-    public void editarDesarrollador(){
+    
+    public Boolean editarDesarrollador(){
+       try {
         Desarrollador desarrollador = desarrolladorService.buscarDesarrolladorPorId(Long.parseLong(tfId.getText()));
         desarrollador.setNombre(tfNombre.getText());
         desarrollador.setPais(tfPais.getText());
-        desarrolladorService.guardarDesarrollador(desarrollador);
-        cargarDatos();
+        return desarrolladorService.guardarDesarrollador(desarrollador);
+       } catch (Exception e) {
+        e.printStackTrace();
+        return false;
+       }
     }
 
     public void eliminarDesarrollador(){
